Updated the hash syntax cop with ruby19_no_mixed_keys style
I've added a new variant to the ruby19 hash style, that behaves the same
as the regular 1.9 style, except when a hash has any keys that cannot be
represented in the new style. In that case, it expects all of the keys
in the hash to be hash rockets, to avoid hashes that mix the two styles
in the same hash.