[Fix #1642] Raise right exception on bad config
In case of syntax errors in the config file, #find_target_files will
raise an exception. Before that exception whould be cought in
`CLI#run`, the `#ensure` block in `Runner#run` would cause *another*
exception to be raised, and only this irrelevant exception would be
caught by `CLI#run` and reported to the user.

Solution: Extract the code with an `#ensure` block to a new method.