Fix false positive for Style/TrailingComma cop
https://github.com/bbatsov/rubocop/issues/1955

The Style/TrailingComma cop reported false positives. Specifically, code
such as `Foo.new({})` and `Foo.new([])` caused false positives when
EnforcedStyleForMultiline was set to 'comma'. The
`multiline?` method, in trailing_comma.rb interpreting the previous
methods as multiline caused the bug. Setting
the EnforcedStyleForMultiline options for the
Style/TrailingComma cop to 'comma' or 'consistent_comma' means multiline
method invocations are expected to have trailing commas after their final
argument. Because `Foo.new({})` is a method invocation, and mistakenly
considered to be multiline, the Style/TrailingComma cop believed it
should be written as `Foo.new({},)`. This is not the intended behavior
of the cop.

* Add checks to the `multiline?` behavior so methods will not
incorrectly be considered multiline.
* Add tests related to the original bugs reported.