[Fix #2006] Fix crash in case of nested offenses
The error happened when `check_alignment` in AutocorrectAlignment adds
an offense with nil node.

```ruby
if offenses.any? { |o| within?(expr, o.location) }
  # If this offense is within a line range that is already being
  # realigned by autocorrect, we report the offense without
  # autocorrecting it. Two rewrites in the same area by the same
  # cop can not be handled. The next iteration will find the
  # offense again and correct it.
  add_offense(nil, expr)
```

In this case `add_offense` passes that nil into `message` method of the
cop that uses AutocorrectAlignment. `FirstParameterIndentation#message`
didn't handle the case when node is nil, so I've added a guard with
a generic message for this case. It looks like the message should be clear enough
since there probably would be a more detailed message reported just before this one.