Keep line breaks in place on WordArray autocorrect
I've found a problem that when I have an array of words that spans
multiple lines

```ruby
array = ['first_word', 'second_word', 'third_word', 'fourth_word',
         'fifth_word', 'sixth_word', 'seveth_word']
```

Autocorrect replaces it with %w() syntax, which is fine, but it also
removes all of the line breaks, which makes the line too long and forces
me to fix it manually.

I've made Syntax/WordArray take into account existing line breaks when
it autocorrects it.