Add Style/LineBreak cops
* `Style/FirstArrayElementLineBreak` checks for a line break before the
  first element in a multi-line array.
* `Style/FirstHashElementLineBreak` checks for a line break before the
  first element in multi-line hash.
* `Style/FirstMethodArgumentLineBreak` checks for a line break before
  the first argument in a multi-line method call.
* `Style/FirstMethodParameterLineBreak` checks for a line break before
  the first parameter in a multi-line method parameter definition.

When large multi-line lists are present in a code base it's often
rather arbitrary to align them based on the first element:

```
long_variable = { foo: 'foo',
                  bar: 'bar',
                  baz: 'baz' }
```

In this example if the variable name is refactored the entire list
formatting also needs to change. This makes diffs, in particular, very
hard to read due to the large number of formatting-only changes:

```
short = { foo: 'foo',
          bar: 'bar',
          baz: 'baz' }
```

These alignments also greatly reduce the available space on a line for
each element, especially if there is a maximum line length enforcement.

This collection of cops along with any appropriate Align or Indent
cop (for example `HashAlign` and `HashIndent`) allows you to enforce a
more diff-friendly and short-line style:

```
variable = {
  foo: 'foo',
  bar: 'bar',
  baz: 'baz' }
```

Care has been taken to ensure that DSL-like method calls are ignored and
do not result in invalid Ruby. I have also added tests for all use-cases
I could find in some large Ruby and Rails code-bases.