Add Style/MultilineAssignmentLayout Cop
This cop checks for a new-line after the assignment operator in
multi-line assignments.

When `EnforcedStyle` is set to `new_line`, this cop enforces the width
efficient style of multi-line assignments. It guarantees that an
assignment in which the right-hand-side spans multiple lines includes
a newline after the assignment operator.

When `EnforcedStyle` is set to `same_line`, this cop guarantees that
an assignment in which the right-hand-side spans multiple lines begins
on the same line as the assignment operator.

* A configurable set of node types, like `if` and `case` are supported.
* Indentation can further by corrected by other cops such as
`Style/IndentationConsistency` and `Style/IndentAssignment`.

https://github.com/bbatsov/ruby-style-guide#indent-conditional-assignment