New source file for Node extensions
Out of the more than 10,000 lines of Ruby code in this project, quite a few lines are
concerned with checking various characteristics of AST nodes. Does this node span more
than one line of source? Does it appear at the beginning of a source line? If it is a
method call, does it have arguments? Is it surrounded by parentheses? And so on...

This kind of logic is duplicated in many places. It would be better to pull out generally
useful query methods and share them. This has already been done to some extent, but it can
be done further.

Since we are working in an OO language, and the project conforms to an OO design style,
when we want to know something about a node, which object should we ask? Ask the Node itself,
of course!

This file is expected to grow quite a bit as duplicated node-query logic is factored out.
In time, it might even need to be split into more than one file.