NodePattern allows extra arguments for predicates/funcalls
The syntax is like this:

    (send equal?(%1) ...)
    (send _x :+ #method(_x, 1, :sym))

The argument lists can contain integer literals, symbol literals, nil,
named wildcards (which must already be bound to a value), and % parameters.

For predicates, which must end with '?' and call methods on the target node,
the arguments will be the only arguments passed to the method. For funcalls,
which are preceded by '#' and which call a method in the context where the
pattern-matching method is defined, the target node is always passed as the
first argument, and any arglist will append extra arguments onto that.