Add Style/IndentAssignment Cop
When assignments span multiple lines, this cop enforces proper
indentation on the right-hand-side of the expression. It only corrects
the first line, as the remaining lines can be corrected by other cops
such as `Style/IndentationConsistency`.

    # bad
    value =
    if foo
      'bar'
    else
      'biz'
    end

    # good
    value =
      if foo # corrected/accepted
      'bar'
    else
      'biz'
    end

I have also added a fix to `Astrolabe::Node#asgn_method_call?`. It used
to consider comparison operators such as `!=` and `===` as assignments -
which they are not. There were also no tests for this method so I've
added them.