Add option AllowHeredoc to Metrics/LineLength
This option can either be `true`, `false`, or a list of heredoc
delimiters. When set to `true`, long lines in all heredocs are allowed.
When set to a list, long lines are only allowed in heredocs delimited by
the delimiters in that list.

The default value of this option is true.

Closes #1407