[Fix #1569] Add `ExpectMatchingDefinition` config param for `Style/FileName`
This has been requested a number of times. It makes `Style/FileName` check that
each source file defines a class or module which matches the file name and path.
For example, if the path is 'lib/foo/bar_baz.rb', the file must define a class
or module called `Foo::BarBaz`.

This option is off by default.