[Fix #2450] Style/Next adjusts indentation when autocorrecting
Until now, `Style/Next` relied on other cops to correct indentation after it
made its own corrections. Now, it adjusts indentation to maintain correct
indentation after a correction (if it was correct before).

This is important, because in some cases, the cops which correct the wrong
indentation may be disabled.

When adjusting indentation, it makes sure not to adjust the bodies of
heredocs.