Return from CLI#run rather than calling Kernel#exit
The return value of `CLI#run` is a UNIX process exit code. However, there were
various places in the code where RC would exit directly, rather than returning
from `CLI#run`.

This is problematic for testing. Yes, we can do this:

    expect { blah }.to exit_with_code(0)

But if someone happens to write a test like this:

    expect(cli.run(args)).to eq 0

...That test can pass *regardless of what the exit code actually was*.

This is because when `Kernel#exit` is called, the current example stops
immediately and *passes*, leaving any remaining assertions in the example
unexecuted.

To avoid such a situation, always raise an exception if execution should
terminate, rather than calling `Kernel#exit`. `CLI#run` can rescue the exception
and return immediately.