[Fix #2546] Handle double rubocop:disable of a cop
When a comment tries to disable a cop that was already disabled
by an earlier comment in the same file, we make it into a
new disabled range. So the given cop is disabled as would be
expected.

The disabled ranges can then be analyzed by Lint/UnneededDisable
and the second disable will be flagged as unnecessary.