[Fix #2253] Add DefaultFormatter config parameter for .rubocop.yml
The new config parameter can be used like:

    AllCops:
      DefaultFormatter: fuubar

It is only applied once, close to the beginning of program execution. Whatever
.rubocop.yml file applies within the working directory from which RuboCop is
run, will decide what the default formatter is.

Previously, setting formatters was the responsibility of Options. This doesn't
work any more. It is not possible to load .rubocop.yml first, extract the value
of DefaultFormatter, and then use it within Options. Why? Because the options
affect how the config file is loaded, from where it is loaded, and so on. We
can't load a config file until we first parse the options.

Therefore, both the selection of a default formatter, as well as adding the
formatter for --auto-gen-config, must be done after `Options#parse`.