[Fix #1895] Remove Rails/DefaultScope cop
Active Record's `default_scope` method can be used in 3 ways:

    default_scope where(a: 1)
    default_scope { where(a: 1) }
    default_scope -> { where(a: 1) }

The `DefaultScope` cop enforced the use of the 2nd form. This is unnecessary,
and causes problems, since in some situations only the 3rd form will do.
(To be specific, in cases where the arguments must be re-evaluated every time
records are retrieved using the default scope, only the 3rd form will work.)

Therefore, remove this cop.