[Fix #2389, #2550] Add TargetRubyVersion config parameter
This parameter can be used to configure which version of the Ruby interpreter
the inspected code is intended to run under. When checking syntax by running
the parser on a snippet of corrected code, RC will use the corresponding parser.
Also, RC will not encourage the use of a construct which is not available under
the target version. Auto-correction will also be careful not to emit code which
will not run on the target interpreter.

Previously, since the parser used was tied to the version of the underlying
Ruby interpreter, not all specs were run on all Ruby versions. Also, it was not
possible to run RC on a 1.9 interpreter to check a project using 2.2 syntax
(for example). Now, all specs run on all Ruby versions, and you can use any
supported interpreter to check any project.

Closes #2389, closes #2550.