Generate --auto-gen-config properly when styles are spread across files
5673c6f broke auto-generation of config, in cases where some files use one style
and other files use another style.

After inspection of each file, RC would effectively "forget" which styles it
had already seen, so if the correct style was seen in one file and an incorrect
style in another file, it would not recognize the conflicting style and
configure the cop as `Enabled: false`.