New Performance/DoubleStartEndWith cop
This cop checks for double `#start_with?` or `#end_with?` calls
separated by `||`. In some cases such calls can be replaced
with an single `#start_with?`/`#end_with?` call.

```ruby
  # bad
  str.start_with?("a") || str.start_with?(Some::CONST)
  str.start_with?("a", "b") || str.start_with?("c")
  var1 = ...
  var2 = ...
  str.end_with?(var1) || str.end_with?(var2)

  # good
  str.start_with?("a", Some::CONST) || str.start_with?(Some::CONST)
  str.start_with?("a", "b", "c")
  var1 = ...
  var2 = ...
  str.end_with?(var1, var2)
```

The corner cases are when arguments of the second `start_with?` call
 have side effects, or when the receiver itself has side effects,
or when the receiver is not a string at all, or when someone
monkey-patched `String#start_with?`