[Fix #2609] Autocorrection works in files which contain __FILE__
Several cops use the AutocorrectUnlessChangingAST module to assist in tricky
autocorrections. This module would fail on any file which contained __FILE__,
since when it tried parsing the corrected source, it did not supply the
original file path to the parser. So the parser would convert __FILE__ to a
completely different string.