[Fix #2586] Remove dependency on Astrolabe
We have a dependency conflict with the `astrolabe` gem. Unfortunately, the
maintainer is not currently accepting pull requests.

Therefore, bring the needed code from `astrolabe` directly into `rubocop`.

This will also make it easy for us to adjust the naming of some methods to
improve code readability. This can be done a bit later.