[Fix #2707] Ignore nested method definitions in Class.new/Module.new
Such method definitions syntactically appear nested, but are in fact
not, since `Class.new` is implemented using `class_eval` and thus the
nested definition defines a method on the new class, not the enclosing
one. In a similar way, `Module.new` is implemented using `module_eval`
and the same applies.