[Fix #2703] Calculate column on first line when BOM is present
Subtract 1 if there's a byte order mark. This is needed
when checking alignment. If we just go by the raw column
number of the range of something on line 1, we will get
column 1 instead of 0 for the first visible character of
the file and conclude that something further down is
misaligned when it's not.