[Fix #2737] Style/GuardClause doesn't flag `elsif`
This cop already had a check for `elsif`, but it did it in a way which didn't
really make sense. In this situation:

    if condition
      return
    elsif other_condition
      something
    end

...it wouldn't report an offense, since the `if` is followed by an `elsif`
clause.

That doesn't make sense, since that code could perfectly well be converted to
use a guard clause. Where we don't want to use a guard clause is:

    if condition
      something
    elsif other_condition
      return
    end