Optimize EmptyLineBetweenDefs
`RuboCop::Node#sibling_index` scans the list of all siblings looking for
the target node, making it O(n) in the number of `def`s in a given
scope. This makes EmptyLineBetweenDefs quadratic in the size of a single
scope.

Instead, intercept `begin` nodes and walk the children explicitly,
keeping track of the previous node without having to do an O(n) scan to
find it.