[Fix #2627] BlockDelimiters accepts {} block call for last value in braceless hash
Previously, the use of `{}` would be flagged in:

    cr.stubs client: mock {
      expects(:email_disabled=).with(true)
      expects :save
    }

However, changing this to use `do...end` would change the meaning of the code,
since it would bind to the outer method call rather than the inner one.

The changes pushed this class over the line count limit, so a couple of tweaks
have been made to save on excess code size.