[Fix #2981] Avoid double trailing comma (#3015)
* [Fix #2981] Do not add double comma when auto-correcting hash arg

For the enforced style `comma`, fix the handling
of a braced hash as last method call argument in
Style/TrailingCommaInArguments. This cop should
not add a comma after the last argument if that
argument has braces that will be removed by
Style/BracesAroundHashParameters.

* Improve test coverage in TrailingComma mixin

This case was missing in the specs.