`Style/StringLiterals` with unicode strings
Now `Style/StringLiterals` works properly for unicode strings,
distinguishing the `España` (which _does not_ need double quotes) from
the `Espa\u00f1a` case (which _does_ need double quotes).

This causes two tests to fail, but I don't think the functionality
provided by those tests was a good thing, so I removed them. If the user
uses unicode control sequences in a string, it's most certainly
intentional and I don't think it's reasonable to detect it as an offense
and autocorrect it to a version that doesn't require double quotes.

It's like the joke...

A chemist walks into a pharmacy and asks the pharmacist, "Do you have
any acetylsalicylic acid?".  "You mean aspirin?", asked the pharmacist.
"That's it, I can never remember that word."

If the user writes "\\x68\\x65\\x6c\\x6c\\x6f" instead of "hello", it's
because she has a reason to do so and RuboCop shouldn't register an
offense for it nor autocorrect it. We might want to add this
simplification as a separate `UnnecessaryControlCharacter` cop, but I
don't think it makes sense here.